pyproject.toml
[project]
name = "homesec"
version = "0.1.1"
description = "Pluggable async home security camera pipeline with detection, VLM analysis, and alerts."
readme = "README.md"
authors = [
    { name = "Lev Neiman", email = "lev.neiman@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["home-security", "computer-vision", "video", "mqtt", "vlm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: Multimedia :: Video",
    "Topic :: Security",
    "Framework :: AsyncIO",
    "Operating System :: OS Independent",
]
dependencies = [
    "dropbox>=12.0.2",
    "fire>=0.7.1",
    "opencv-python>=4.12.0.88",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "ultralytics>=8.3.226",
    "pyyaml>=6.0.3",
    "aiohttp>=3.13.2",
    "anyio>=4.0.0",
    "greenlet>=3.3.0",
    "pyftpdlib>=2.1.0",
]

[project.scripts]
homesec = "homesec.cli:main"

[project.urls]
Homepage = "https://github.com/lan17/homesec"
Source = "https://github.com/lan17/homesec"
Issues = "https://github.com/lan17/homesec/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/homesec"]

# External plugins should use this entry point group:
# [project.entry-points."homesec.plugins"]
# my_package = "my_package"  # Point to package containing decorated plugins
#
# Example external plugin package structure:
# my_package/__init__.py should contain:
#   from homesec.plugins.filters import filter_plugin
#   @filter_plugin(name="custom")
#   def custom_plugin(): ...

[dependency-groups]
dev = [
    "asyncpg-stubs>=0.31.1",
    "ipykernel>=7.1.0",
    "mypy>=1.19.1",
    "nbstripout>=0.8.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
